Avoid warnings.

+++ tools/ppm2tiff.c
@@ -1,4 +1,4 @@
-/* $Id: ppm2tiff.c,v 1.7 2004-09-02 14:51:04 dron Exp $ */
+/* $Id: ppm2tiff.c,v 1.8 2004-09-21 12:36:02 dron Exp $ */
 
 /*
  * Copyright (c) 1991-1997 Sam Leffler
@@ -223,7 +223,7 @@ processCompressOptions(char* opt)
 		compression = COMPRESSION_PACKBITS;
 	else if (strneq(opt, "jpeg", 4)) {
 		char* cp = strchr(opt, ':');
-		if (cp && isdigit(cp[1]))
+		if (cp && isdigit((int)cp[1]))
 			quality = atoi(cp+1);
 		if (cp && strchr(cp, 'r'))
 			jpegcolormode = JPEGCOLORMODE_RAW;
